Revamp the DIscordRPC category
